+++ react/admin/ApprovedReviewsTable.tsx
@@ -64,7 +64,9 @@ export const ApprovedReviewsTable: FC = () => {
     } = review
 
     return {
-      date: reviewDateTime,
+      date: `${intl.formatDate(reviewDateTime)} ${intl.formatTime(
+        reviewDateTime
+      )}`,
       product: {
         productId,
         sku,